Move more Charge and Customer tests over
